fix: rename dataviz / dataVisualization to datavis / dataVisualisation
BREAKING CHANGE: `@udir-design/theme/dataviz.css` is renamed to `@udir-design/theme/datavis.css`
