+++ @udir-design/react/.storybook/style.css
@@ -8,7 +8,7 @@
 */
 @import '../../css/src/external.css';
 @import '../../css/src/icons.css';
-@import '../../theme/dist/dataviz.css';
+@import '../../theme/dist/datavis.css';
 
 body {
   font-family: 'Inter', sans-serif;
