+++ @udir-design/react/demo-pages/dashboard-demo/tabs/overview/Overview.tsx
@@ -3,7 +3,7 @@ import { HighchartsReact } from 'highcharts-react-official';
 import { useRef, useState } from 'react';
 import { Table } from 'src/components/table';
 import { Heading } from 'src/components/typography/heading/Heading';
-import { getHighchartsTheme } from 'src/utilities/dataVisualization';
+import { getHighchartsTheme } from 'src/utilities/datavis';
 import { DataControls } from '../../components/data-controls/DataControls';
 import { LoadChart } from '../../components/loading/LoadChart';
 import { TabStructure } from '../../components/tab-structure/TabStructure';
