+++ @udir-design/react/package.json
@@ -30,9 +30,9 @@
       "import": "./dist/utilities/form/alpha.js",
       "require": "./dist/utilities/form/alpha.cjs"
     },
-    "./utilities/dataVisualization/alpha": {
-      "import": "./dist/utilities/dataVisualization/alpha.js",
-      "require": "./dist/utilities/dataVisualization/alpha.cjs"
+    "./utilities/datavis/alpha": {
+      "import": "./dist/utilities/datavis/alpha.js",
+      "require": "./dist/utilities/datavis/alpha.cjs"
     },
     "./html": {
       "types": "./dist/html.d.ts"
