+++ @udir-design/react/src/design-tokens/Colors.mdx
@@ -12,7 +12,7 @@ import {
   DatavisSeqColorDisplay,
   DatavisSeqDivColorDisplay,
   DatavisDivGradient,
-} from './docs/color/dataviz-color-display';
+} from './docs/color/datavis-color-display';
 
 <Meta title="design-tokens/Farger" />
 <Unstyled>
