--- @udir-design/react/src/utilities/dataVisualization/index.ts
@@ -1,2 +1,0 @@
-export * from './dataVisualization';
-export * from './highcharts';
