+++ @udir-design/react/src/utilities/datavis/index.ts
@@ -1,0 +1,2 @@
+export * from './dataVisualisation';
+export * from './highcharts';
