+++ @udir-design/react/typedoc.json
@@ -2,7 +2,7 @@
   "cleanOutputDir": true,
   "displayBasePath": "./src",
   "entryPoints": [
-    "./src/utilities/dataVisualization/alpha.ts",
+    "./src/utilities/datavis/alpha.ts",
     "./src/utilities/hooks/alpha.ts",
     "./src/utilities/hooks/beta.ts",
     "./src/utilities/form/alpha.ts"
