+++ @udir-design/react/vite.config.ts
@@ -88,8 +88,7 @@ export default defineConfig({
         alpha: 'src/alpha.ts',
         beta: 'src/beta.ts',
         'utilities/form/alpha': 'src/utilities/form/alpha.ts',
-        'utilities/dataVisualization/alpha':
-          'src/utilities/dataVisualization/alpha.ts',
+        'utilities/datavis/alpha': 'src/utilities/datavis/alpha.ts',
       },
       name: '@udir-design/react',
       // Change this to the formats you want to support.
