+++ @udir-design/theme/package.json
@@ -16,7 +16,7 @@
       "types": "./dist/types.d.ts",
       "default": "./dist/index.css"
     },
-    "./dataviz.css": "./dist/dataviz.css"
+    "./datavis.css": "./dist/datavis.css"
   },
   "files": [
     "./dist"
