+++ test-apps/nextjs/src/app/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import '@udir-design/react/style.css';
-import '@udir-design/theme/dataviz.css';
+import '@udir-design/theme/datavis.css';
 import './global.css';
 import { TestFooter as Footer } from './components/TestFooter';
 import { TestHeader as Header } from './components/TestHeader';
