+++ test-apps/vite/tsconfig.app.json
@@ -10,8 +10,8 @@
       "src/utilities/form/*": [
         "./node_modules/@udir-design/react/dist/utilities/form/alpha"
       ],
-      "src/utilities/dataVisualization": [
-        "./node_modules/@udir-design/react/dist/utilities/dataVisualization/alpha"
+      "src/utilities/datavis": [
+        "./node_modules/@udir-design/react/dist/utilities/datavis/alpha"
       ],
       "src/utilities/*": ["./node_modules/@udir-design/react/dist/alpha"]
     },
